+++ packages/eas-cli/src/analytics/commands/common.ts
@@ -1,0 +1,24 @@
+import { Analytics, Event } from './events';
+
+export type TrackingContext = Record<string, string | number | boolean>;
+
+export async function withAnalyticsAsync<Result>(
+  fn: () => Promise<Result>,
+  analytics: {
+    successEvent: Event;
+    failureEvent: Event;
+    trackingCtx: TrackingContext;
+  }
+): Promise<Result> {
+  try {
+    const result = await fn();
+    Analytics.logEvent(analytics.successEvent, analytics.trackingCtx);
+    return result;
+  } catch (error: any) {
+    Analytics.logEvent(analytics.failureEvent, {
+      ...analytics.trackingCtx,
+      reason: error.message,
+    });
+    throw error;
+  }
+}
