+++ packages/eas-cli/src/analytics/rudderstackClient.ts
@@ -3,8 +3,8 @@ import os from 'os';
 import { URL } from 'url';
 import { v4 as uuidv4 } from 'uuid';
 
-import UserSettings from './user/UserSettings';
-import { easCliVersion } from './utils/easCli';
+import UserSettings from '../user/UserSettings';
+import { easCliVersion } from '../utils/easCli';
 
 const PLATFORM_TO_ANALYTICS_PLATFORM: { [platform: string]: string } = {
   darwin: 'Mac',
