+++ packages/eas-cli/src/submit/android/AndroidSubmitter.ts
@@ -1,6 +1,7 @@
 import { Platform } from '@expo/eas-build-job';
 import chalk from 'chalk';
 
+import { SubmissionEvent } from '../../analytics/commands/events';
 import {
   AndroidSubmissionConfigInput,
   SubmissionAndroidReleaseStatus,
@@ -8,7 +9,6 @@ import {
   SubmissionFragment,
 } from '../../graphql/generated';
 import { SubmissionMutation } from '../../graphql/mutations/SubmissionMutation';
-import { SubmissionEvent } from '../../utils/analytics';
 import formatFields from '../../utils/formatFields';
 import { Archive, ArchiveSource, getArchiveAsync } from '../ArchiveSource';
 import BaseSubmitter, { SubmissionInput } from '../BaseSubmitter';
